Short solution for order book update slow
Order book update slow will make our system thought the order has not been fulfilled, which will lead to not update holdingStatus. This will be an issue since next time, we will ignore this pair for buying/selling.

Easy fix will be to check if cancel order has been successful. If it is not, then in most cases, the order has already been fulfilled.